mtree(8): Fix crc() prototype.
It is taken from usr.bin/cksum/crc.c. This fixes a -Wlto-type-mismatch
warning reported by zrj. Similar to the fix for ckdist(1) as pushed in
e69bb3d9c4c6d6073c9c9c4a54753458fae55c7e.

+++ usr.sbin/mtree/extern.h
@@ -40,7 +40,7 @@ extern uint32_t crc_total;
 extern u_int keys;
 
 int	 compare(NODE *, FTSENT *);
-int	 crc(int, u_long *, u_long *);
+int	 crc(int, uint32_t *, off_t *);
 void	 cwalk(void);
 char	*flags_to_string(u_long);
